// src/types/mesurement-record.ts
/**
 * @license
 * Copyright 2019 Google Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Cdn } from './cdn';
import { Connectivity } from './connectivity';
import { Sdk } from './sdk';

export interface MeasurementRecord {
  test_run: number;
  version: string;
  sdk: Sdk;
}

export interface NetworkLatencyRecord extends MeasurementRecord {
  cdn: Cdn;
  connectivity: Connectivity;
  device: string;
  browser_version: string;
  ttfb_ms: number;
  download_ms: number;
}

export interface ExecutionLatencyRecord extends MeasurementRecord {
  device: string;
  browser_version: string;
  parse_ms: number;
  exec_ms: number;
}

export interface BinarySizeRecord extends MeasurementRecord {
  size_byte: number;
}